Initial commit
The Send Logs to Loki GitHub Action collects logs from all jobs in a GitHub Actions workflow and sends them to a Loki instance.
Logs are labeled with metadata like job names, job IDs, and custom labels, making them easily searchable and organized in Loki.

The code within this repository was migrated from the https://github.com/skroutz/github-actions/ repository (commit d02eb0b55668f10d7cd03a73b9359206c32311fa) .
This migration was undertaken to facilitate a restructuring of the code, aligning it with established best practices, with the ultimate goal of publishing it on the GitHub Actions Marketplace.